Move index.html and add reading list page
The index.html file was moved to the public folder for better organization. Also, a reading list page was created as a new feature, with a link added to it in the menu section of the index.html page. This will help users easily navigate to the reading list from the main blog page.